refactor(vulnerable-code): Prepare for multiple API versions
Move the existing API v1 implementation to a dedicated `AdviceProvider` and
let `VulnerableCode` delegate to the provider selected by the configured
API version.

Split tests so plugin-level behavior and API v1 behavior are covered
separately. Also structure the functional test by API version to prepare
for adding API v3 coverage.
