+++ plugins/advisors/vulnerable-code/src/main/kotlin/VulnerableCodeConfiguration.kt
@@ -24,11 +24,19 @@ import org.ossreviewtoolkit.plugins.api.OrtPluginOption
 import org.ossreviewtoolkit.plugins.api.Secret
 
 /**
+ * The supported VulnerableCode API versions.
+ */
+enum class VulnerableCodeApiVersion {
+    V1
+}
+
+/**
  * The configuration for VulnerableCode as security vulnerability provider.
  */
 data class VulnerableCodeConfiguration(
     /**
-     * The base URL of the VulnerableCode REST API. By default, the public VulnerableCode instance is used.
+     * The base URL of the VulnerableCode REST API. By default, the public VulnerableCode instance is used. The
+     * implementation will take care of normalizing the API URL to contain the "/api/" part.
      */
     @OrtPluginOption(defaultValue = VULNERABLE_CODE_BASE_URL)
     val serverUrl: String,
@@ -41,5 +49,11 @@ data class VulnerableCodeConfiguration(
     /**
      * The read timeout for the server connection in seconds. Defaults to whatever is the HTTP client's default value.
      */
-    val readTimeout: Long?
+    val readTimeout: Long?,
+
+    /**
+     * The VulnerableCode API version to use.
+     */
+    @OrtPluginOption(defaultValue = "V1")
+    val apiVersion: VulnerableCodeApiVersion
 )
