+++ plugins/advisors/vulnerable-code/src/test/kotlin/TestUtils.kt
@@ -19,8 +19,11 @@
 
 package org.ossreviewtoolkit.plugins.advisors.vulnerablecode
 
+import com.github.tomakehurst.wiremock.WireMockServer
+
 import io.kotest.core.TestConfiguration
 
+import org.ossreviewtoolkit.model.AdvisorDetails
 import org.ossreviewtoolkit.model.Identifier
 import org.ossreviewtoolkit.model.OrtResult
 import org.ossreviewtoolkit.model.Package
@@ -70,3 +73,21 @@ internal fun generatePackagesRequest(purls: Collection<String> = packages): Stri
  * Generate the JSON body of the request to query vulnerability information about the [Package] with the given [id].
  */
 internal fun generatePackagesRequest(id: Identifier): String = generatePackagesRequest(listOf(id.toPurl()))
+
+/**
+ * The advisor details expected in results produced by the VulnerableCode test instances.
+ */
+internal val details = AdvisorDetails(VulnerableCodeFactory.descriptor.id)
+
+/**
+ * Create a configuration for the VulnerableCode implementation with the given [apiVersion] that points to the local
+ * [server]. If [apiUrl] is true, append the legacy "/api/" suffix to the server URL.
+ */
+internal fun createConfig(
+    server: WireMockServer,
+    apiVersion: VulnerableCodeApiVersion,
+    apiUrl: Boolean = false
+): VulnerableCodeConfiguration {
+    val url = "http://localhost:${server.port()}" + if (apiUrl) "/api/" else ""
+    return VulnerableCodeConfiguration(url, null, null, apiVersion)
+}
